requirements.txt
# Pipeline HR Analytics - Dépendances Python 3.10
# ===================================================

# Manipulation de données
pandas>=1.5.0,<2.1.0
numpy>=1.24.0,<1.25.0
faker>=18.0.0,<20.0.0
openpyxl>=3.0.10,<3.2.0
xlsxwriter>=3.0.9,<3.2.0

# Streaming et messagerie
kafka-python>=2.0.2,<2.1.0
confluent-kafka>=2.1.0,<2.4.0

# AWS et stockage cloud
boto3>=1.26.0,<1.35.0
s3fs>=2023.6.0,<2024.7.0
awscli>=1.29.0,<1.34.0

# Apache Spark pour analytics et ML (compatible Python 3.10)
pyspark>=3.4.0,<3.5.0
scikit-learn>=1.3.0,<1.4.0
mlflow>=2.5.0,<2.10.0

# Visualisation et reporting
matplotlib>=3.6.0,<3.8.0
seaborn>=0.12.0,<0.13.0
plotly>=5.15.0,<5.18.0

# API et web services
fastapi>=0.100.0,<0.105.0
uvicorn>=0.22.0,<0.25.0
requests>=2.28.0,<2.32.0

# Configuration et logging
python-dotenv>=1.0.0,<1.1.0
pyyaml>=6.0.0,<6.1.0
loguru>=0.7.0,<0.8.0

# Base de données et connecteurs
sqlalchemy>=1.4.0,<2.1.0
psycopg2-binary>=2.9.5,<2.10.0

# Tests et qualité de code
pytest>=7.2.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<5.14.0

# Sécurité et monitoring
cryptography>=40.0.0,<42.0.0
prometheus-client>=0.16.0,<0.20.0

# Utilitaires Python 3.10
pytz>=2023.3,<2025.0
tqdm>=4.64.0,<4.67.0
schedule>=1.2.0,<1.3.0